Modulo Mensualidad

+++ Conservatorio.DATOS/PagoMatricula.cs
@@ -7,7 +7,7 @@ namespace Conservatorio.DATOS
         public virtual int IdPagoMatricula { get; set; }
         public virtual string Periodo { get; set; }
         public virtual int Año { get; set; }
-        public virtual decimal Monto { get; set; }
+        public virtual double Monto { get; set; }
         public virtual DateTime FechaPago { get; set; }
         public virtual Estudiante Estudiante { get; set; }
         public virtual string Comentario { get; set; }
